worker/requirements.txt
nodriver
pymongo
pyyaml
pika
requests
aiofiles
redis